+++ homebox/__init__.py
@@ -1,6 +1,6 @@
 """A Python wrapper for the Homebox API."""
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 
 from .client import HomeboxClient